move changes to beta.py
